pyproject.toml
[tool.poetry]
name = "0923"
version = "0.1.0"
description = ""
authors = ["Replit user <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
python-binance = "^1.0.16"
requests = "^2.28.1"
pandas = "^1.5.0"
matplotlib = "^3.6.0"
numpy = "^1.23.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
